refactor: extract the code generatore from the manipulators, few comments and new lines to code generated for asp readability
